[EI-155] [api] rename: refactoring에 따른 폴더 구조 및 파일명 변경
