--- api/src/numerical-guidance/application/command/insert-indicator-ticker/insert-indicator-ticker.command.ts
@@ -1,9 +1,0 @@
-import { ICommand } from '@nestjs/cqrs';
-
-export class InsertIndicatorTickerCommand implements ICommand {
-  constructor(
-    readonly id: string,
-    readonly ticker: string,
-    readonly type: string,
-  ) {}
-}
