+++ api/src/numerical-guidance/application/command/update-indicator-board-metadata-name/update-indicator-board-metadata-name.command.handler.ts
@@ -1,9 +1,9 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { Inject, Injectable } from '@nestjs/common';
 import { UpdateIndicatorBoardMetadataNameCommand } from './update-indicator-board-metadata-name.command';
-import { LoadIndicatorBoardMetadataPort } from '../../port/persistence/load-indiactor-board-metadata.port';
+import { LoadIndicatorBoardMetadataPort } from '../../port/persistence/indicator-board-metadata/load-indiactor-board-metadata.port';
 import { Transactional } from 'typeorm-transactional';
-import { UpdateIndicatorBoardMetadataNamePort } from '../../port/persistence/update-indicator-board-metadata-name.port';
+import { UpdateIndicatorBoardMetadataNamePort } from '../../port/persistence/indicator-board-metadata/update-indicator-board-metadata-name.port';
 import { IndicatorBoardMetadata } from '../../../domain/indicator-board-metadata';
 
 @Injectable()
