--- api/src/numerical-guidance/application/port/indicator-list/load-indicator-list.port.ts
@@ -1,5 +1,0 @@
-import { IndicatorListDto } from '../../query/get-indicator-list/indicator-list.dto';
-
-export interface LoadIndicatorListPort {
-  loadIndicatorList(): Promise<IndicatorListDto>;
-}
