+++ api/src/numerical-guidance/application/port/persistence/indicator-board-metadata/create-indicator-board-metadata.port.ts
@@ -1,4 +1,4 @@
-import { IndicatorBoardMetadata } from '../../../domain/indicator-board-metadata';
+import { IndicatorBoardMetadata } from '../../../../domain/indicator-board-metadata';
 
 export interface CreateIndicatorBoardMetadataPort {
   createIndicatorBoardMetaData(indicatorBoardMetaData: IndicatorBoardMetadata, memberId: number): Promise<string>;
