+++ api/src/numerical-guidance/application/port/persistence/indicator-board-metadata/delete-indicator-board-metadata.port.ts
