+++ api/src/numerical-guidance/application/port/persistence/indicator-board-metadata/update-indicator-board-metadata-name.port.ts
@@ -1,4 +1,4 @@
-import { IndicatorBoardMetadata } from '../../../domain/indicator-board-metadata';
+import { IndicatorBoardMetadata } from '../../../../domain/indicator-board-metadata';
 
 export interface UpdateIndicatorBoardMetadataNamePort {
   updateIndicatorBoardMetadataName(indicatorBoardMetadata: IndicatorBoardMetadata): Promise<void>;
