--- api/src/numerical-guidance/application/query/get-indicator-list/get-indicator-list.query.ts
@@ -1,5 +1,0 @@
-import { IQuery } from '@nestjs/cqrs';
-
-export class GetIndicatorListQuery implements IQuery {
-  constructor() {}
-}
