+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator/entity/indicator.entity.ts
